# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BasicAuth protects our entity endpoints.",
        "title": "Assessment Tax API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/admin/deductions/personal": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "description": " ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tax.updateTaxDeductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/tax.updatePersonalDeductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError-array_validator_Field"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError-string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError-string"
                        }
                    }
                }
            }
        },
        "/tax/calculations": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tax"
                ],
                "parameters": [
                    {
                        "description": " ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tax.taxCalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/tax.taxCalculateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError-array_validator_Field"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError-string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError-string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.ResponseError-array_validator_Field": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validator.Field"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "http.ResponseError-string": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "errors": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "tax.allowance": {
            "type": "object",
            "required": [
                "allowanceType"
            ],
            "properties": {
                "allowanceType": {
                    "type": "string",
                    "example": "donation"
                },
                "amount": {
                    "type": "number",
                    "minimum": 0,
                    "example": 200000
                }
            }
        },
        "tax.taxCalculateRequest": {
            "type": "object",
            "required": [
                "allowances"
            ],
            "properties": {
                "allowances": {
                    "type": "array",
                    "uniqueItems": true,
                    "items": {
                        "$ref": "#/definitions/tax.allowance"
                    }
                },
                "totalIncome": {
                    "type": "number",
                    "minimum": 0,
                    "example": 500000
                },
                "wht": {
                    "type": "number",
                    "minimum": 0,
                    "example": 25000
                }
            }
        },
        "tax.taxCalculateResponse": {
            "type": "object",
            "properties": {
                "tax": {
                    "type": "number"
                },
                "taxLevel": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tax.taxLevel"
                    }
                },
                "taxRefund": {
                    "type": "number"
                }
            }
        },
        "tax.taxLevel": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "string"
                },
                "tax": {
                    "type": "number"
                }
            }
        },
        "tax.updatePersonalDeductResponse": {
            "type": "object",
            "properties": {
                "personalDeduction": {
                    "type": "number"
                }
            }
        },
        "tax.updateTaxDeductRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
        "validator.Field": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}